added session user in todos page and [id] page and finished the section

+++ devtalles_/03-admin-todo/src/app/api/todos/route.ts
@@ -1,4 +1,5 @@
 import prisma from '@/app/lib/prisma'
+import { getUserSessionServer } from '@/auth/actions/authActions';
 import { NextResponse } from 'next/server'
 import * as yup from 'yup';
 
@@ -36,10 +37,16 @@ const postSchema = yup.object({
 
 export async function POST(request: Request) {
 
+  const user = await getUserSessionServer();
+  if (!user) {
+    return NextResponse.json('No autorize', { status: 401 })
+  }
+
   try {
     const body = await postSchema.validate( await request.json() );
+    const { complete, description } = body;
     
-    const todo = await prisma.todo.create({ data: body })
+    const todo = await prisma.todo.create({ data: {complete, description, userId: user.id} })
   
     return NextResponse.json(todo)
   } catch (err) {
@@ -49,8 +56,13 @@ export async function POST(request: Request) {
 
 export async function DELETE(request: Request) {
 
+  const user = await getUserSessionServer(); // para que solo guarde los todos del usuario en concreto
+  if (!user) {
+    return NextResponse.json('No autorize', { status: 401 })
+  }
+
   try {
-    await prisma.todo.deleteMany({ where: { complete: true }})
+    await prisma.todo.deleteMany({ where: { complete: true, userId: user.id }})
   
     return NextResponse.json('good deleted')
   } catch (err) {
